Modularize css and js structures
