+++ src/main.module.css
@@ -2,6 +2,7 @@
   flex-grow: 0;
   flex-shrink: 0;
   flex-basis: 50%;
+  padding: 1em;
 }
 
 .animationContainer{
@@ -17,6 +18,8 @@
   animation-delay: 1s;
   animation-duration: 1s;
   animation-fill-mode: forwards;
+  background-size: cover;
+  background-position: 50% 50%;
 }
 
 @keyframes slide-image{
